+++ bmds/plotting.py
@@ -1,5 +1,6 @@
 import matplotlib as mpl
-import matplotlib.pyplot as plt
+mpl.use('Agg')  # prevent matplotlib framework issues
+import matplotlib.pyplot as plt  # noqa
 
 __all__ = []