+++ tests/test_execution.py
@@ -6,6 +6,7 @@ import bmds
 from bmds import models
 from simple_settings import settings
 from simple_settings.utils import settings_stub
+import pytest
 
 from .fixtures import *  # noqa
 
@@ -22,7 +23,6 @@ def test_executable_path():
         if inspect.isclass(obj):
             if obj not in parents and issubclass(obj, parents):
                 exe = obj.get_exe_path()
-                print(obj.__name__, exe)
                 assert os.path.exists(exe)
 
 
@@ -261,6 +261,7 @@ def test_bad_datasets(bad_cdataset, bad_ddataset):
         assert np.isclose(total_time, timeout) or total_time > timeout
 
 
+@pytest.mark.skip(reason='timeout failure is unreliable')
 def test_timeout():
     # confirm that timeout setting works as expected; slow dataset provided
     # by CEBS team on 2017-11-12. Only slow in BMDS version 2.6; fixed in 2.7.